Consolidate traits
