--- src/compiler/typescript_to_rust/azle_generate/src/ts_ast/traits/get_name.rs
@@ -1,3 +1,0 @@
-pub trait GetName {
-    fn get_name(&self) -> &str;
-}
