--- src/compiler/typescript_to_rust/azle_generate/src/ts_ast/traits/get_source_info.rs
@@ -1,6 +1,0 @@
-pub trait GetSourceInfo {
-    fn get_source(&self) -> String;
-    fn get_line_number(&self) -> usize;
-    fn get_origin(&self) -> String;
-    fn get_range(&self) -> (usize, usize);
-}
