--- src/compiler/typescript_to_rust/azle_generate/src/ts_ast/traits/to_display_string.rs
@@ -1,3 +1,0 @@
-pub trait GetSourceText {
-    fn get_source_text(&self) -> String;
-}
